fix: #63 코드 개선 및 파일 구조화
